Solution to the water container problem in Ruby
This solution to the water container problem is written in Ruby rather than Python. It also uses different logic. Instead of using a nested loop, it uses a two pointer system that starts at either end of the list. This should be O(n) rather than O(n^2) as before.